# back-end/swagger.yaml
openapi: 3.0.3
info:
  title: APIs tisco - OpenAPI 3.0
  description: APIs for tisco
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: http://159.138.242.188:3000/api/v1
...
components:
  securitySchemes:
    beareAuth:
        type: http
        scheme: bearer
        name: Authorization
        in: header
        bearerFormat: JWT
        description: >-
          Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"
...
  schemas:
    Account:
      type: object
      properties:
        hasPassword:
          type: boolean
        accId:
          type: integer
        uuidAccount:
          type: string
        compCode:
          type: string
        condiVersion:
          type: string
        condiAccept:
          type: integer
        accPicture:
          type: string
          nullable: true
        accFirstname:
          type: string
        accLastname:
          type: string
          nullable: true
        active:
          type: boolean
        createdBy:
          type: string
        updatedBy:
          type: string
        createdDatetime:
          type: string
          format: date-time
        updatedDatetime:
          type: string
          format: date-time
        chanelLogin:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              uuidAccount:
                type: string
              loginPlatform:
                type: string
              loginData:
                type: string
              loginVerify:
                type: integer
              loginSecret:
                type: string
                nullable: true
              lineClientId:
                type: string
                nullable: true
              targetConsentCode:
                type: string
                nullable: true
              targetConsentVersion:
                type: string
                nullable: true
              acceptStatus:
                type: integer
              status:
                type: string
              isDefault:
                type: boolean
              isLineLiff:
                type: boolean
              active:
                type: boolean
              lastLogin:
                type: string
                format: date-time
              createdBy:
                type: string
              updatedBy:
                type: string
              createdDatetime:
                type: string
                format: date-time
              updatedDatetime:
                type: string
                format: date-time
        peronaIcon:
          type: object
          properties:
            iconId:
              type: integer
            iconLevel:
              type: integer
            iconName:
              type: string
            iconImage:
              type: object
              properties:
                size:
                  type: integer
                type:
                  type: string
                object:
                  type: string
                fileUrl:
                  type: string
                newName:
                  type: string
                originalName:
                  type: string
            iconDetail:
              type: string
            iconMinScore:
              type: integer
            iconMaxScore:
              type: integer
            active:
              type: boolean
            createdBy:
              type: string
            updatedBy:
              type: string
            createdDatetime:
              type: string
              format: date-time
            updatedDatetime:
              type: string
              format: date-time
paths:
  /accounts:
    get:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Get profile information"
      description: "Retrieves profile information of the currently logged-in user."
      responses:
        '200':
          description: "Profile information retrieved successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
            application/xml:
              schema:
                $ref: '#/components/schemas/Account'
  /accounts/update-profile:
    put:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Update account information"
      description: "Updates the profile information of the currently logged-in user."
      parameters:
        - name: "accPicture"
          description: "Account picture URL"
          required: false
          schema:
            type: "string"
        - name: "accFirstname"
          description: "Account first name"
          required: false
          schema:
            type: "string"
        - name: "compCode"
          description: "Company code"
          required: false
          schema:
            type: "string"
      responses:
        200:
          description: "Account information updated successfully"
  /accounts/lock-password:
    patch:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Lock password"
      description: "Locks the password of the currently logged-in user."
      responses:
        200:
          description: "Password locked successfully"
  /accounts/forgot-pwd:
    put:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Save forgot password"
      description: "Saves the new password for the currently logged-in user."
      parameters:
        - name: "accPassword"
          description: "New account password"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: "Password updated successfully"
  /accounts/code-company:
    put:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Change code company"
      description: "Changes the company code associated with the currently logged-in user."
      parameters:
        - name: "code"
          description: "New company code"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: "Company code changed successfully"
  /accounts/disconnect:
    post:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Save disconnected channel login"
      description: "Saves the disconnected channel login information for the currently logged-in user."
      parameters:
        - name: "chanelId"
          description: "Channel ID to disconnect"
          required: true
          schema:
            type: "integer"
      responses:
        200:
          description: "Channel login disconnected successfully"
  /accounts/connect:
    post:
      tags:
        - account
      security:
        - beareAuth: []
      summary: "Save connected channel login"
      description: "Saves the connected channel login information for the currently logged-in user."
      parameters:
        - name: "loginData"
          description: "Login data for the new channel"
          required: true
          schema:
            type: "object"
        - name: "loginPlatform"
          description: "Login platform for the new channel"
          required: true
          schema:
            type: "string"
            enum: ["EMAIL", "PHONE", "LINE"]
      responses:
        200:
          description: "Channel login connected successfully"
...